lab6/service_app/requirements.txt
fastapi==0.68.1
pydantic==1.8.2
uvicorn==0.15.0
python-jose==3.3.0
python-multipart==0.0.5
kafka-python==2.0.2
redis==4.3.4
sqlalchemy==1.4.41
psycopg2-binary>=2.9.1
httpx==0.23.0
PyJWT>=2.3.0